Suggest more relevant recommendations
Previously, a reader was recommended a random article after
each read. This may not always be desirable because the reader
would most likely be encouraged to read more articles if the
recommendation is related to the current post. In this commit,
an article with tags similar to the current article is
recommendeded. If no articles with matching tags are found, a
random article from the entire collection is selected.