Redirect to admin page after the authentication modal has closed
The "on login" callback is triggered each time a logged in user navigates
to the home page, redirecting them to the admin page. This prevents
authenticated users from accessing the home page, which is not the desired
behavior. Using the "on close" callback in stead ensures that redirection
only happens if an authentication modal was accessed on the home page, which
currently only happens during password resets and first-time registration.